Struct libvda::encode::Session

pub struct Session { /* fields omitted */ }

Represents an encode session.

Implementations

impl Session

pub fn pipe(&self) -> &File

Returns a reference for the pipe that notifies of encode events.

pub fn read_event(&mut self) -> Result<Event>

Reads an Event object from a pipe provided by an encode session.

pub fn encode(
&self,
input_buffer_id: VeaInputBufferId,
fd: BufferFd,
planes: &[FramePlane],
timestamp: i64,
force_keyframe: bool
) -> Result<()>

Sends an encode request for an input buffer given as fd with planes described by planes. The timestamp of the frame to encode is typically provided in milliseconds by timestamp. force_keyframe` indicates to the encoder that the frame should be encoded as a keyframe.

When the input buffer has been filled, an EncoderEvent::ProcessedInputBuffer event can be read from the event pipe.

The caller is responsible for passing in a unique value for input_buffer_id which can be referenced when the event is received.

fd will be closed after encoding has occurred.

pub fn use_output_buffer(
&self,
output_buffer_id: VeaOutputBufferId,
fd: BufferFd,
offset: u32,
size: u32
) -> Result<()>

Provides a buffer for storing encoded output.

When the output buffer has been filled, an EncoderEvent::ProcessedOutputBuffer event can be read from the event pipe.

The caller is responsible for passing in a unique value for output_buffer_id which can be referenced when the event is received.

This function takes ownership of fd.

pub fn request_encoding_params_change(
&self,
bitrate: Bitrate,
framerate: u32
) -> Result<()>

Requests encoding parameter changes.

The request is not guaranteed to be honored by libvda and could be ignored by the backing encoder implementation.

pub fn flush(&self) -> Result<()>

Flushes the encode session.

When this operation has completed, Event::FlushResponse can be read from the event pipe.